Added base html pages to facilitate spotify log in flow
