+++ src/html/dashboard.html
@@ -1,0 +1,9 @@
+<html>
+  <head></head>
+  <body>
+    <h1>Dashboard</h1>
+    <div>
+      Manage your playlists here.
+    </div>
+  </body>
+</html>
