+++ src/html/index.html
@@ -1,0 +1,11 @@
+<html>
+  <head></head>
+  <body>
+    <div>
+      <div id="login">
+        <h1>First, log in to spotify</h1>
+        <a href="/login">Log in</a>
+      </div>
+    </div>
+  </body>
+</html>
